% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{CleanMetabolonData}
\alias{CleanMetabolonData}
\title{Cleans the untidy dataset given the relevent sample, metabolite and count
Data Frames. These Frames have fields specific to the DILT1D xls output from Metabolon}
\usage{
CleanMetabolonData(sampleInfo, metaboliteInfo, metaboliteData)
}
\arguments{
\item{sampleInfo}{A Data Frame of sample metadata}

\item{metaboliteInfo}{A Data Frame of metabolite metadata}

\item{metaboliteData}{A Data Frame of count data}
}
\value{
A list of 4 Data Frames with the sample and metabolite metadata and the
Wide and Long versions of the Metabolite Count data
}
\description{
Cleans the untidy dataset given the relevent sample, metabolite and count
Data Frames. These Frames have fields specific to the DILT1D xls output from Metabolon
}

