% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{fetchClinchemAnalyte}
\alias{fetchClinchemAnalyte}
\title{returns a particular Clinical Chemistry analyte from the ClinChem dataframe for subjects in the DILT1D trial}
\usage{
fetchClinchemAnalyte(clinChem, analyte)
}
\arguments{
\item{clinChem}{A Data Frame of the ClinChem measurements for each subject per visit}
}
\value{
A character vector of the analyto to be returned
}
\description{
returns a particular Clinical Chemistry analyte from the ClinChem dataframe for subjects in the DILT1D trial
}

