% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{LoadDILT1DCovariates}
\alias{LoadDILT1DCovariates}
\title{Loads the covariate data from the ipswich network share for DILT1D}
\usage{
LoadDILT1DCovariates(covariatesFileStem, covariatesDate, covariatesSourceFile)
}
\arguments{
\item{covariatesFileStem}{A filestem on the ipswich share that documents covariates for the
DILT1D trial participants}

\item{covariatesDate}{The date for the covariates file for which to choose the covariates}

\item{covariatesSourceFile}{An R file of functions to be sourced to retrieve the data, also on ipswich}
}
\value{
A Data Frame of covariate data for the DILT1D trial
}
\description{
Loads the covariate data from the ipswich network share for DILT1D
}
\examples{
covariateData <- LoadDILT1DCovariates(covariatesFile = "/ipswich/path/to/file-stem-", covariatesDate = 'YYYY-MM-DD' covariatesFile = "/ipswich/path/to/file.R")
}

