% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{LoadDILT1DDataVolNormalised}
\alias{LoadDILT1DDataVolNormalised}
\title{Loads the DILT1D dataset given the relevent sample, metabolite and count file
Cleans the files which have fields specific to the DILT1D output from Metabolon}
\usage{
LoadDILT1DDataVolNormalised(sampleFile, metaboliteFile, dataFile)
}
\arguments{
\item{sampleFile}{A character filepath to the sample metadata}

\item{metaboliteFile}{A character filepath to the metabolite metadata}

\item{dataFile}{A character filepath to the count data}
}
\value{
A list of Data Frames with the sample and metabolite metadata and the
volume normalised count data (but not further normalised)
}
\description{
Loads the DILT1D dataset given the relevent sample, metabolite and count file
Cleans the files which have fields specific to the DILT1D output from Metabolon
}

