% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{PlotAg15Panel}
\alias{PlotAg15Panel}
\title{Plots a presentation ready plot representing the correlations between Ag1,5 and
point in DILT1D}
\usage{
PlotAg15Panel(sampleInfo, covariateData, data15AG, gluData, hbaData,
  cpeptideData)
}
\arguments{
\item{sampleInfo}{A Data Frame containing sample data for DILT1D Metabolon data}

\item{covariateData}{A Data Frame containing covariate data for DILT1D}

\item{data15AG}{A Data Frame containing Metabolon Ag1,5 expression data}

\item{gluData}{A Data Frame containing clinical glucose measurements for DILT1D patients}

\item{hbaData}{A Data Frame containing clinical HbA1c measurements for DILT1D patients}

\item{cpeptideData}{A Data Frame containing clinical C-peptide measurements for DILT1D patients}
}
\value{
A list of ggplots containing correlations between Ag1,5 and the analytes
}
\description{
Plots a presentation ready plot representing the correlations between Ag1,5 and
point in DILT1D
}
\examples{

}

