% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{PlotAnalyteAvailability}
\alias{PlotAnalyteAvailability}
\title{Plots a presentation ready plot representing the number of sampled analytes available at each time
point in DILT1D}
\usage{
PlotAnalyteAvailability(sampleInfo, covariates, cpeptideData, hbaData, agData,
  gluData, smbgData)
}
\arguments{
\item{sampleInfo}{A Data Frame containing sample data for DILT1D Metabolon data}

\item{cpeptideData}{A Data Frame containing C-Peptide expression data}

\item{hbaData}{A Data Frame containing HbA1c expression data}

\item{agData}{A Data Frame containing Metabolon Ag1,5 expression data}

\item{gluData}{A Data Frame containing clinical glucose measurements}

\item{smbgData}{A Data Frame containing clinical smbg measurements}

\item{covariateData}{A Data Frame containing covariate data for DILT1D}
}
\value{
A ggplot containing a tiled plot of the data availability
}
\description{
Plots a presentation ready plot representing the number of sampled analytes available at each time
point in DILT1D
}
\examples{

}

