% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{PlotCovariates}
\alias{PlotCovariates}
\title{Plots the covariates of sex, age and dose in DILT1D study}
\usage{
PlotCovariates(dilt1dCovariates)
}
\arguments{
\item{dilt1dCovariates}{a Data Frame of the DILT1D covariates from the doses file}
}
\value{
A ggplot of the covariates faceted by sex
}
\description{
Plots the covariates of sex, age and dose in DILT1D study
}

