% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{PlotGlucosePanel}
\alias{PlotGlucosePanel}
\title{Plots a presentation ready plot representing the correlations between Ag1,5 and
point in DILT1D}
\usage{
PlotGlucosePanel(sampleInfo, covariateData, metabolonGlu, gluData, smbgData)
}
\arguments{
\item{sampleInfo}{A Data Frame containing sample data for DILT1D Metabolon data}

\item{covariateData}{A Data Frame containing covariate data for DILT1D}

\item{metabolonGlu}{A Data Frame containing Metabolon glucose expression data}

\item{gluData}{A Data Frame containing clinical glucose measurements for DILT1D patients}

\item{smbgData}{A Data Frame containing SMBG measurements for DILT1D patients}
}
\value{
A list of ggplots containing correlations between the various glucose measurements
}
\description{
Plots a presentation ready plot representing the correlations between Ag1,5 and
point in DILT1D
}
\examples{

}

