% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{PlotMetaboliteVarianceComponents}
\alias{PlotMetaboliteVarianceComponents}
\title{Plots a presentation ready plot representing the variance components of the
metabolite data in DILT1D}
\usage{
PlotMetaboliteVarianceComponents(ICCs)
}
\arguments{
\item{ICCs}{A Data Frame containing variance components data from the Linear Mixed Model}
}
\value{
A list of ggplots containing plots of the technical, within subject and between subject variation
}
\description{
Plots a presentation ready plot representing the variance components of the
metabolite data in DILT1D
}
\examples{

}

