% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{PlotSexMetaboliteComparison}
\alias{PlotSexMetaboliteComparison}
\title{Plots a presentation ready plot comparing the effect sizes of the Krumsiek and DILT1D
sex lined metabolites}
\usage{
PlotSexMetaboliteComparison(dilt1dCoeffs, dilt1dPVals, krumResults,
  metaboliteInfo)
}
\arguments{
\item{dilt1dCoeffs}{A Data Frame containing coefficents for the sex metabolite regressions}

\item{metaboliteInfo}{A Data Frame containing the metabolite metadata}

\item{dilt1dCoeffs}{A Data Frame containing p-values for the sex metabolite regressions}

\item{dilt1dCoeffs}{A Data Frame literature results from Krumsiek sex metabolite regressions}
}
\value{
A 2 element List. A ggplot containing the effect sizes of the sex metabolites in Krumsiek and DILT1D
and the data frame used for the plotting
}
\description{
Plots a presentation ready plot comparing the effect sizes of the Krumsiek and DILT1D
sex lined metabolites
}
\examples{
p <- PlotSexMetaboliteComparison(dilt1dCoeffs, dilt1dPVals, krumResults)

}

