% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{PlotVolcano}
\alias{PlotVolcano}
\title{Plots a presentation ready plot comparing the effect sizes of the Krumsiek and DILT1D
sex lined metabolites}
\usage{
PlotVolcano(DFregression, regressor, cMetInfo, plotLabels = TRUE,
  plotColour = TRUE, correction = "fdr")
}
\arguments{
\item{sexRegressionResults}{A Data Frame containing cleaned information about the samples}
}
\value{
A ggplot volcano plot containing the log2 fold change vs -log10P value of the fold changes across Visits
in the DILT1D data trial
}
\description{
Plots a presentation ready plot comparing the effect sizes of the Krumsiek and DILT1D
sex lined metabolites
}
\examples{
p <- PlotSexMetaboliteComparison(sexRegressionResults)

}

