% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{readDILT1DClinchemData}
\alias{readDILT1DClinchemData}
\title{returns the Clinical Chemistry measurements (ClinChem) for subjects in the DILT1D trial}
\usage{
readDILT1DClinchemData(clinChemFile)
}
\arguments{
\item{smbgFile}{A character file location for the ClinChem measurements}
}
\value{
A Data Frame of the ClinChem measurements for each subject per visit
}
\description{
returns the Clinical Chemistry measurements (ClinChem) for subjects in the DILT1D trial
}

