% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{readDILT1DCpeptideData}
\alias{readDILT1DCpeptideData}
\title{returns the CPeptide measurements for subjects in the DILT1D trial}
\usage{
readDILT1DCpeptideData(cPeptideFile)
}
\arguments{
\item{cPeptideFile}{A character file location for the CPeptide measurements}
}
\value{
A Data Frame of the C-peptide measurements for each subject per visit
}
\description{
returns the CPeptide measurements for subjects in the DILT1D trial
}

