% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{readDILT1DHBAData}
\alias{readDILT1DHBAData}
\title{returns the HbA1c measurements for subjects in the DILT1D trial}
\usage{
readDILT1DHBAData(hbaFile)
}
\arguments{
\item{hbaFile}{A character file location for the HbA1c measurements}
}
\value{
A Data Frame of the HbA1c measurements for each subject per visit
}
\description{
returns the HbA1c measurements for subjects in the DILT1D trial
}

