% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{readDILT1DSMBGData}
\alias{readDILT1DSMBGData}
\title{returns the Self Measured Blood Glucose (SMBG) measurements for subjects in the DILT1D trial}
\usage{
readDILT1DSMBGData(smbgFile)
}
\arguments{
\item{smbgFile}{A character file location for the SMBG measurements}
}
\value{
A Data Frame of the SMBG measurements for each subject per visit
}
\description{
returns the Self Measured Blood Glucose (SMBG) measurements for subjects in the DILT1D trial
}

