% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regressions.R
\name{RegressFoldchange}
\alias{RegressFoldchange}
\title{performs univariate fold change regressions for each metabolite between two time points in DILT1D study
\deqn{\Delta M_i = X\Beta _i + \Epsilon _i}}
\usage{
RegressFoldchange(covariates, cSampleInfo, cMetDataLong,
  timepoints = c("V0Pre", "V0Post"), predictor = "DIFF")
}
\arguments{
\item{covariates}{The DILT1D covariate data frame}

\item{cSampleInfo}{The DILT1D dataframe for sample metadata}

\item{cMetDataLong}{The DILT1D dataframe for metabolite counts (or z-scores)}

\item{timepoints}{A vector of visits with which to calculate the fold changes}

\item{the}{predictor variable for the regression}
}
\value{
A list of coefficents, standard errors, t-values, p-values and the models for the regressions
}
\description{
performs univariate fold change regressions for each metabolite between two time points in DILT1D study
\deqn{\Delta M_i = X\Beta _i + \Epsilon _i}
}

