% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{RunAnalyteAnalysis}
\alias{RunAnalyteAnalysis}
\title{Runs the DILT1D analyte analysis for comparing Ag1,5}
\usage{
RunAnalyteAnalysis(fileLocationsList, outputDir)
}
\arguments{
\item{fileLocationsList}{A list of names file locations}

\item{outputDir}{A character filepath store the output from the analysis}
}
\description{
The method reads in the data, cleans it, and estimates
the correlations between CPeptide, Ag1,5, clinical glucose
and Self-Measured glucose in the DILT1D trial
}

