% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{RunQCAnalysis}
\alias{RunQCAnalysis}
\title{Performs A comparison of QC methods on Metabolon data}
\usage{
RunQCAnalysis(fileLocationsList, outputDir)
}
\arguments{
\item{fileLocationsList}{A list of named file locations}

\item{outputDir}{A character filepath store the output from the analysis}
}
\description{
The method reads in the data, cleans it, and estimates
the impact of different QC methods (NOMIS, CCMN and median)
on a pca and positive control (glucose and sex metabolites)
}

