% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{RunUniVariateAnalysis}
\alias{RunUniVariateAnalysis}
\title{Runs the DILT1D univariate analysis across visits}
\usage{
RunUniVariateAnalysis(fileLocationsList, outputDir)
}
\arguments{
\item{fileLocationsList}{A list of named file locations}

\item{outputDir}{A character filepath store the output from the analysis}
}
\description{
The method reads in the data, cleans it, and estimates
a univariate model for each metabolite \deqn{y = X\Beta + \Epsilon}
between the V0Pre visit and subsequent timepoints. The output
is placed in a directory specified by the user
}

