% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{RunVarianceComponentsAnalysis}
\alias{RunVarianceComponentsAnalysis}
\title{Performs the variance components analysis across DILT1D dataset}
\usage{
RunVarianceComponentsAnalysis(fileLocationsList, outputDir)
}
\arguments{
\item{fileLocationsList}{A list of named file locations}

\item{outputDir}{A character filepath store the output from the analysis}
}
\description{
The method reads in the data, cleans it, and estimates
a variance components model for the DILT1D dataset using
visits V0Pre, V4, V5 assuming drug washout
The output is placed in a directory specified by the user
}

