% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{SexMetabolitesFromKrumsiek}
\alias{SexMetabolitesFromKrumsiek}
\title{Loads in a list of metabolites that are significantly differentially
expressed between sexes. This list is a result of a mWAS analysis
by Krumsiek et al}
\usage{
SexMetabolitesFromKrumsiek(sexMappingFile = "Data/KrumsiekMapping.csv",
  field = "RAW", cMetInfo, cMetDataLong)
}
\arguments{
\item{sexMappingFile}{A character filename that documents a mapping between the
metabolites in the Krumsiek panel (3rd generation) and the Metabolites in DILT1D
(4th generation)}

\item{cMetInfo}{A Data Frame of metabolite metadata}

\item{cMetDataLong}{A Data Frame of metabolite counts}
}
\value{
A list of three Data Frames; the metabolite info of the sex metabolites;
the data for the sex metabolites in long format and the Krumsief lookup which has
the summary statistics from the analysis
}
\description{
Loads in a list of metabolites that are significantly differentially
expressed between sexes. This list is a result of a mWAS analysis
by Krumsiek et al
}

